refactor
merge ChargePointServiceConfigUpdater and ChargePointConfigUpdater
