+++ src/main/java/de/rwth/idsg/steve/service/ChargePointConfigUpdater.java
@@ -19,10 +19,17 @@
 package de.rwth.idsg.steve.service;
 
 import de.rwth.idsg.steve.SteveException;
+import de.rwth.idsg.steve.ocpp.OcppProtocol;
+import de.rwth.idsg.steve.ocpp.OcppSecurityProfile;
+import de.rwth.idsg.steve.ocpp.OcppVersion;
+import de.rwth.idsg.steve.repository.dto.ChargePointSelect;
 import de.rwth.idsg.steve.service.notification.OcppStationSecurityBasicAuthChanged;
 import de.rwth.idsg.steve.service.notification.OcppStationSecurityProfileChanged;
+import de.rwth.idsg.steve.service.notification.OcppStationWebSocketConnected;
 import de.rwth.idsg.steve.web.dto.RestCallback;
 import de.rwth.idsg.steve.web.dto.ocpp.ChangeConfigurationParams;
+import de.rwth.idsg.steve.web.dto.ocpp.ConfigurationKeyEnum;
+import de.rwth.idsg.steve.web.dto.ocpp.GetConfigurationParams;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import ocpp.cp._2015._10.ConfigurationStatus;
@@ -37,9 +44,44 @@ import static de.rwth.idsg.steve.web.dto.ocpp.ConfigurationKeyEnum.SecurityProfi
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ChargePointServiceConfigUpdater {
+public class ChargePointConfigUpdater {
 
+    private final ChargePointService chargePointService;
+
+    // use this when we need to wait for the response to process it
     private final OcppOperationsService ocppOperationsService;
+    // use this when fire-and-forget: no need to wait for or to process the response
+    private final ChargePointServiceClient chargePointServiceClient;
+
+    @EventListener
+    public void getAndUpdateCpoName(OcppStationWebSocketConnected notification) {
+        try {
+            var ocppVersion = notification.getOcppVersion();
+            if (ocppVersion == OcppVersion.V_12 || ocppVersion == OcppVersion.V_15) {
+                return;
+            }
+
+            var chargeBoxId = notification.getChargeBoxId();
+            var registration = chargePointService.getRegistrationDirect(chargeBoxId);
+            if (registration.isEmpty()) {
+                return;
+            }
+
+            // CpoName came with ocpp 1.6 security extension, which also introduced the security profiles. profile 0
+            // represents the old world before the security extension. since it is not relevant, skip this task.
+            if (registration.get().securityProfile() == OcppSecurityProfile.Profile_0) {
+                return;
+            }
+
+            // Send request to get CpoName. Default impl of GetConfigurationTask takes care of updating the database.
+            var params = new GetConfigurationParams();
+            params.setChargePointSelectList(List.of(new ChargePointSelect(OcppProtocol.V_16_JSON, chargeBoxId)));
+            params.setConfKeyList(List.of(ConfigurationKeyEnum.CpoName.name()));
+            chargePointServiceClient.getConfiguration(params);
+        } catch (Exception e) {
+            log.error("Failed", e);
+        }
+    }
 
     @EventListener
     public void tryToChangeSecurityProfileAtStation(OcppStationSecurityProfileChanged event) {
